Add a `composer.json` file
... making the exiting (development) dependencies explicit and easier to find for anyone looking for them.

Includes validating the `composer.json` file on the Travis builds against the lowest and highest stable PHP versions supported by Composer as well as the plugin.

Includes adding the `composer.lock` file to the `.gitignore` list.
Until there are non-dev dependencies which need to be locked to specific versions, there is no need to commit the `composer.lock` file.